Add support for TF-IDF for unigrams
The bot will no longer use unigrams in the same way it used to.
It used to look at 1-grams (unigrams), 2-grams, 3-grams and
4-grams in the same way, but now it only looks at 2-grams, 3-grams
and 4-grams in that way. Unigrams get special treatment now.

For unigrams (that is, single words), the TF-IDF is computed.
The corpus used is all the training sentences, and the "document"
is the training sententences for the Intent being checked. So the
unigrams get a "weight" which is equal to its TF-IDF value, and
these weights are added to the values that are calculated in the
old way for the 2-grams, 3-grams and 4-grams.